Change Log 58
Updated a whole bunch o' stuff just so that the enemy could be properly
tested.
It can now move and chose a random direction to move in (but it stays
stuck in collisions at the moment)
Will attempt to update a better version so it can be easily applied to
other enemies.